+++ test/components/businesses/open_graph_tags_component_test.rb
@@ -7,7 +7,7 @@ class OpenGraphTagsComponentTest < ViewComponent::TestCase
     render_inline OpenGraphTagsComponent.new
 
     assert_meta property: "og:type"
-    assert_meta property: "og:title", content: "railsdevs"
+    assert_meta property: "og:title", content: "RailsDevs"
     assert_meta property: "og:description", content_begin_with: "The reverse job board"
     assert_meta property: "og:url"
     assert_meta property: "twitter:card"
@@ -19,7 +19,7 @@ class OpenGraphTagsComponentTest < ViewComponent::TestCase
       description: "And a custom description."
     )
 
-    assert_meta property: "og:title", content: "Custom title · railsdevs"
+    assert_meta property: "og:title", content: "Custom title · RailsDevs"
     assert_meta property: "og:description", content: "And a custom description."
   end