+++ test/components/developers/open_graph_meta_tags_component_test.rb
@@ -9,7 +9,7 @@ module Developers
 
       render_inline OpenGraphTagsComponent.new(developer:)
 
-      assert_meta property: "og:title", content: "hero text · railsdevs"
+      assert_meta property: "og:title", content: "hero text · RailsDevs"
       assert_meta property: "og:description", content: "bio text"
       assert_meta property: "og:url", content: "http://test.host/developers/123"
     end